requirements.txt
Django~=5.0.4
djangorestframework~=3.15.1
requests~=2.31.0
django-model-utils~=4.5.0
